+++ src/relations/Serves.java
@@ -3,29 +3,29 @@ package relations;
 public class Serves {
     private String CarPlate;
     private int OrderID;
-    private int TimeStart;
-    private int TimeFinish;
+	private long TimeStart;
+	private long TimeFinish;
 
-    public Serves(String carPlate, int orderID, int timeStart, int timeFinish) {
+	public Serves(String carPlate, int orderID, long timeStart, long timeFinish) {
         CarPlate = carPlate;
         OrderID = orderID;
         TimeStart = timeStart;
         TimeFinish = timeFinish;
     }
 
-    public int getTimeFinish() {
+	public long getTimeFinish() {
         return TimeFinish;
     }
 
-    public void setTimeFinish(int timeFinish) {
+	public void setTimeFinish(long timeFinish) {
         TimeFinish = timeFinish;
     }
 
-    public int getTimeStart() {
+	public long getTimeStart() {
         return TimeStart;
     }
 
-    public void setTimeStart(int timeStart) {
+	public void setTimeStart(long timeStart) {
         TimeStart = timeStart;
     }
